refactor: restructure rootFolder code
